requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
sqlalchemy~=2.0.41
sqlalchemy[asyncio]~=2.0.41
asyncpg~=0.30.0
psycopg2~=2.9.10
PyYAML~=6.0.2
pydantic~=2.11.5
pydantic[email]~=2.11.5
passlib~=1.7.4
PyJWT~=2.10.1
cryptography~=45.0.3
bcrypt~=4.3.0
aiobotocore~=2.23.0
botocore~=1.38.27
pillow~=11.2.1
python-multipart~=0.0.20
